+++ webots_ros2_abb/test/test_flake8.py
@@ -21,5 +21,5 @@ import pytest
 @pytest.mark.flake8
 @pytest.mark.linter
 def test_flake8():
-    rc = main(argv=[])
+    rc = main(argv=['--linelength', '128'])
     assert rc == 0, 'Found errors'