+++ webots_ros2_core/webots_ros2_core/webots_differential_drive_node.py
@@ -12,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import argparse
 import sys
+
 from math import cos, sin
 import rclpy
 from rclpy.time import Time
@@ -23,6 +23,7 @@ from geometry_msgs.msg import Twist, TransformStamped
 from tf2_ros import TransformBroadcaster
 from webots_ros2_core.webots_node import WebotsNode
 from webots_ros2_core.math_utils import euler_to_quaternion
+from webots_ros2_core.utils import get_node_name_from_args
 
 
 class WebotsDifferentialDriveNode(WebotsNode):
@@ -89,7 +90,6 @@ class WebotsDifferentialDriveNode(WebotsNode):
 
         # Initialize timer
         self._last_odometry_sample_time = self.robot.getTime()
-        self.create_timer(self.timestep / 1000, self._publish_odometry_data)
 
     def _cmd_vel_callback(self, twist):
         self.get_logger().info('Message received')
@@ -100,16 +100,21 @@ class WebotsDifferentialDriveNode(WebotsNode):
         self.left_motor.setVelocity(left_omega)
         self.right_motor.setVelocity(right_omega)
 
-    def _publish_odometry_data(self):
+    def step(self, ms):
+        super().step(ms)
+
         stamp = Time(seconds=self.robot.getTime()).to_msg()
 
         time_diff_s = self.robot.getTime() - self._last_odometry_sample_time
         left_wheel_ticks = self.left_wheel_sensor.getValue()
         right_wheel_ticks = self.right_wheel_sensor.getValue()
 
+        if time_diff_s == 0.0:
+            return
+
         # Calculate velocities
-        v_left_rad = (left_wheel_ticks - self._prev_left_wheel_ticks) / time_diff_s if time_diff_s > 0.0 else 0.0
-        v_right_rad = (right_wheel_ticks - self._prev_right_wheel_ticks) / time_diff_s if time_diff_s > 0.0 else 0.0
+        v_left_rad = (left_wheel_ticks - self._prev_left_wheel_ticks) / time_diff_s
+        v_right_rad = (right_wheel_ticks - self._prev_right_wheel_ticks) / time_diff_s
         v_left = v_left_rad * self._wheel_radius
         v_right = v_right_rad * self._wheel_radius
         v = (v_left + v_right) / 2
@@ -193,11 +198,9 @@ class WebotsDifferentialDriveNode(WebotsNode):
 def main(args=None):
     rclpy.init(args=args)
 
-    parser = argparse.ArgumentParser()
-    parser.add_argument('--name', default='diff_driver', help='Name of your diff drive node')
-    args, _ = parser.parse_known_args()
-
-    differential_drive = WebotsDifferentialDriveNode(args.name, args=args)
+    webots_robot_name = get_node_name_from_args()
+    differential_drive = WebotsDifferentialDriveNode(webots_robot_name, args=args)
+    differential_drive.start_device_manager()
     rclpy.spin(differential_drive)
     rclpy.shutdown()