+++ webots_ros2_core/webots_ros2_core/webots_node.py
@@ -19,13 +19,18 @@
 import argparse
 import os
 import sys
+from rclpy.time import Time
+from webots_ros2_core.joint_state_publisher import JointStatePublisher
+from webots_ros2_core.devices.device_manager import DeviceManager
 
-from webots_ros2_core.utils import append_webots_python_lib_to_path
+from webots_ros2_core.utils import append_webots_python_lib_to_path, get_node_name_from_args
+from webots_ros2_core.tf_publisher import TfPublisher
 
 from webots_ros2_msgs.srv import SetInt
 
 from rosgraph_msgs.msg import Clock
 
+import rclpy
 from rclpy.node import Node
 
 try:
@@ -37,10 +42,10 @@ except Exception as e:
 
 
 class WebotsNode(Node):
-
-    def __init__(self, name, args=None):
+    def __init__(self, name, args=None, enableTfPublisher=False, enableJointState=False):
         super().__init__(name)
         self.declare_parameter('synchronization', False)
+        self.declare_parameter('use_joint_state_publisher', False)
         parser = argparse.ArgumentParser()
         parser.add_argument('--webots-robot-name', dest='webotsRobotName', default='',
                             help='Specifies the "name" field of the robot in Webots.')
@@ -56,8 +61,21 @@ class WebotsNode(Node):
         self.timer = self.create_timer(timer_period, self.timer_callback)
         self.sec = 0
         self.nanosec = 0
+        self.__device_manager = None
+        if enableTfPublisher:
+            if self.robot.getSupervisor():
+                self.tfPublisher = TfPublisher(self.robot, self)
+            else:
+                self.get_logger().warn('Impossible to publish transforms because the "supervisor"'
+                                       ' field is false.')
+        if self.get_parameter('use_joint_state_publisher').value or enableJointState:
+            self.jointStatePublisher = JointStatePublisher(self.robot, '', self)
 
     def step(self, ms):
+        if self.get_parameter('use_joint_state_publisher').value:
+            self.jointStatePublisher.publish()
+        if self.__device_manager:
+            self.__device_manager.step()
         if self.robot is None or self.get_parameter('synchronization').value:
             return
         # Robot step
@@ -66,14 +84,8 @@ class WebotsNode(Node):
             self.robot = None
             sys.exit(0)
         # Update time
-        time = self.robot.getTime()
-        self.sec = int(time)
-        # rounding prevents precision issues that can cause problems with ROS timers
-        self.nanosec = int(round(1000 * (time - self.sec)) * 1.0e+6)
-        # Publish clock
         msg = Clock()
-        msg.clock.sec = self.sec
-        msg.clock.nanosec = self.nanosec
+        msg.clock = Time(seconds=self.robot.getTime()).to_msg()
         self.clockPublisher.publish(msg)
 
     def timer_callback(self):
@@ -83,3 +95,33 @@ class WebotsNode(Node):
         self.step(request.value)
         response.success = True
         return response
+
+    def start_device_manager(self, config=None):
+        """
+        Start automatic ROSification of Webots devices available in the robot.
+
+        Kwargs:
+            config (dict): Dictionary of properties in format::
+
+                {
+                    [device_name]: {
+                        [property_name]: [property_value]
+                    }
+                }
+
+        """
+        self.__device_manager = DeviceManager(self, config)
+
+
+def main(args=None):
+    rclpy.init(args=args)
+
+    webots_robot_name = get_node_name_from_args()
+    driver = WebotsNode(webots_robot_name, args=args)
+    driver.start_device_manager()
+    rclpy.spin(driver)
+    rclpy.shutdown()
+
+
+if __name__ == '__main__':
+    main()