+++ webots_ros2/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package webots_ros2
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.0.0 (2020-09-01)
+------------------
+* Improved support for macOS
+
 0.0.3 (2020-06-15)
 ------------------
 * Updated to Webots R2020a