+++ webots_ros2_epuck/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package webots_ros2_epuck
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.0.0 (2020-09-01)
+------------------
+* Use the webots_ros2_core::WebotsDifferentialDriveNode class
+
 0.0.3 (2020-06-15)
 ------------------
 * Initial version