+++ webots_ros2_importer/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package webots_ros2_importer
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.0.0 (2020-09-01)
+------------------
+* Added the '--multi-file', '--static-base', '--tool-slot' and '--rotation' arguments.
+
 0.0.3 (2020-06-15)
 ------------------
 * Initial version